uefi: fix fwts_uefi_rt_support_status_get()
The UEFI 2.8 Errata A specification has clarified that
RuntimeServicesSupported is not a UEFI variable but a value in the
EFI_RT_PROPERTIES_TABLE configuration table. Since Linux 5.11 the value can
be retrieved via an IOCTL call.

Replace the code trying to read the non-existent RuntimeServicesSupported
variable by the IOCTL call.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>